requirements.txt
asgiref==3.7.2
Django==4.2.3
gunicorn==21.2.0
numpy==1.25.0
opencv-python==4.8.0.74
packaging==23.1
Pillow==10.0.0
pypng==0.20220715.0
PyQRCode==1.2.1
qrcode==7.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
